Refactor backtesting screen associated files
