+++ src/components/main_screens/home/invest/job/BuyingPowerAndHoldings.jsx
