+++ src/components/main_screens/home/invest/job/JobScreen.jsx
@@ -12,19 +12,19 @@ import {
   View,
 } from "react-native";
 import { Snackbar } from "react-native-paper";
-import { TRADE_HISTORY_FETCH_AMOUNT } from "../../../../constants/ApiConstants";
-import AlgoquantApiContext from "../../../../constants/ApiContext";
-import { CHIP_JOB_TYPES } from "../../../../constants/ChipJobTypeEnum";
-import { THEME } from "../../../../constants/Theme";
-import { timeframeEnums } from "../../../../constants/graphEnums";
-import { stopJobModalBuilder } from "../../../../helpers/modalFactory";
-import { snackbarCleanUp } from "../../../../helpers/snackbarCleanup";
-import { jobHistoryColumns } from "../../../../helpers/tableColumns";
-import BuyingPowerAndHoldings from "../../../reusable_components/BuyingPowerAndHoldings";
-import CustomGraph from "../../../reusable_components/CustomGraph";
-import CustomModal from "../../../reusable_components/CustomModal";
-import CustomTable from "../../../reusable_components/CustomTable";
-import GraphDetailsHeader from "../../../reusable_components/GraphDetailsHeader";
+import { TRADE_HISTORY_FETCH_AMOUNT } from "../../../../../constants/ApiConstants";
+import AlgoquantApiContext from "../../../../../constants/ApiContext";
+import { CHIP_JOB_TYPES } from "../../../../../constants/ChipJobTypeEnum";
+import { THEME } from "../../../../../constants/Theme";
+import { timeframeEnums } from "../../../../../constants/graphEnums";
+import { stopJobModalBuilder } from "../../../../../helpers/modalFactory";
+import { snackbarCleanUp } from "../../../../../helpers/snackbarCleanup";
+import { jobHistoryColumns } from "../../../../../helpers/tableColumns";
+import CustomGraph from "../../../../reusable_components/CustomGraph";
+import CustomModal from "../../../../reusable_components/CustomModal";
+import CustomTable from "../../../../reusable_components/CustomTable";
+import GraphDetailsHeader from "../../../../reusable_components/GraphDetailsHeader";
+import BuyingPowerAndHoldings from "./BuyingPowerAndHoldings";
 
 export default function JobScreen(props) {
   const { jobID, jobType } = props.route.params;
